+++ services/inference/tsfminference/inference_handler.py
@@ -2,7 +2,6 @@
 
 import datetime
 import logging
-from abc import abstractmethod
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
@@ -11,17 +10,15 @@ import pandas as pd
 from .inference_payloads import (
     BaseMetadataInput,
     BaseParameters,
-    ForecastingMetadataInput,
-    ForecastingParameters,
     PredictOutput,
 )
-from .service_handler import ForecastingServiceHandler, HandlerFunction, ServiceHandlerBase
+from .service_handler import HandlerFunction, ServiceHandler
 
 
 LOGGER = logging.getLogger(__file__)
 
 
-class InferenceHandler(ServiceHandlerBase):
+class InferenceHandler(ServiceHandler):
     @classmethod
     def load(
         cls, model_id: str, model_path: Union[str, Path]
@@ -65,9 +62,9 @@ class InferenceHandler(ServiceHandlerBase):
             return None, RuntimeError("Service wrapper has not yet been prepared; run `handler.prepare()` first.")
 
         try:
-            result = self._run(data, schema=schema, parameters=parameters, **kwargs)
+            result = self.implementation.run(data, schema=schema, parameters=parameters, **kwargs)
             encoded_result = encode_dataframe(result)
-            counts = self._calculate_data_point_counts(
+            counts = self.implementation.calculate_data_point_counts(
                 data, output_data=result, schema=schema, parameters=parameters, **kwargs
             )
             return PredictOutput(
@@ -79,77 +76,6 @@ class InferenceHandler(ServiceHandlerBase):
 
         except Exception as e:
             return None, e
-
-    @abstractmethod
-    def _run(
-        self,
-        data: pd.DataFrame,
-        schema: Optional[BaseMetadataInput] = None,
-        parameters: Optional[BaseParameters] = None,
-        **kwargs,
-    ) -> pd.DataFrame:
-        """Abstract method for run to be implemented by model owner in derived class"""
-        ...
-
-    @abstractmethod
-    def _calculate_data_point_counts(
-        self,
-        data: pd.DataFrame,
-        output_data: Optional[pd.DataFrame] = None,
-        schema: Optional[BaseMetadataInput] = None,
-        parameters: Optional[BaseParameters] = None,
-    ) -> Dict[str, int]:
-        """Abstract method for counting datapoints in input and output implemented by model owner in derived class"""
-        ...
-
-
-class ForecastingInferenceHandler(ForecastingServiceHandler, InferenceHandler):
-    def run(
-        self,
-        data: pd.DataFrame,
-        future_data: Optional[pd.DataFrame] = None,
-        schema: Optional[ForecastingMetadataInput] = None,
-        parameters: Optional[ForecastingParameters] = None,
-        **kwargs,
-    ) -> Tuple[PredictOutput, None] | Tuple[None, Exception]:
-        """Perform an inference request
-
-        Args:
-            data (pd.DataFrame): A pandas dataframe containing historical data.
-            future_data (Optional[pd.DataFrame], optional): A pandas dataframe containing future data. Defaults to None.
-            schema (Optional[ForecastingMetadataInput], optional): Service request schema. Defaults to None.
-            parameters (Optional[ForecastingParameters], optional): Service requst parameters. Defaults to None.
-
-        Returns:
-            Tuple[PredictOutput, None] | Tuple[None, Exception]: If successful, returns a tuple containing a PredictionOutput
-                object as the first element. If unsuccessful, a tuple with an exception as the second element will be returned.
-        """
-
-        return super().run(data, future_data=future_data, schema=schema, parameters=parameters, **kwargs)
-
-    @abstractmethod
-    def _run(
-        self,
-        data: pd.DataFrame,
-        future_data: Optional[pd.DataFrame] = None,
-        schema: Optional[ForecastingMetadataInput] = None,
-        parameters: Optional[ForecastingParameters] = None,
-        **kwargs,
-    ) -> pd.DataFrame:
-        """Abstract method for run to be implemented by model owner in derived class"""
-        ...
-
-    @abstractmethod
-    def _calculate_data_point_counts(
-        self,
-        data: pd.DataFrame,
-        future_data: Optional[pd.DataFrame] = None,
-        output_data: Optional[pd.DataFrame] = None,
-        schema: Optional[ForecastingMetadataInput] = None,
-        parameters: Optional[ForecastingParameters] = None,
-    ) -> Dict[str, int]:
-        """Abstract method for counting datapoints in input and output implemented by model owner in derived class"""
-        ...
 
 
 def encode_dataframe(result: pd.DataFrame, timestamp_column: str = None) -> Dict[str, List[Any]]:
